cambio un cosito de int a float

+++ Shared_Libraries/commons_propias/commons_propias/commons_propias.h
@@ -64,7 +64,7 @@ typedef struct {
 	int id;
 	estado estado;
 	pthread_t hilo;
-	int retardo;
+	float retardo;
 	int ultimaRafaga;
 	float ultimaEstimacion;
 	int socket;
